refactor: organize android wallet and models
