+++ android/app/src/main/java/org/cashu/wallet/Core/Wallet/WalletDatabaseRecovery.kt
@@ -1,0 +1,10 @@
+package org.cashu.wallet.Core
+
+internal fun shouldAttemptWalletDatabaseRecovery(error: Throwable): Boolean {
+    val normalized = (error.message ?: error.toString()).lowercase()
+    return normalized.contains("sqlite") ||
+        normalized.contains("database") ||
+        normalized.contains("corrupt") ||
+        normalized.contains("malformed") ||
+        normalized.contains("walletdb")
+}
