+++ android/app/src/main/java/org/cashu/wallet/Core/Wallet/WalletManager.kt
@@ -1,6 +1,5 @@
 package org.cashu.wallet.Core
 
-import java.net.HttpURLConnection
 import java.net.URL
 import java.util.UUID
 import kotlinx.coroutines.CoroutineExceptionHandler
@@ -13,10 +12,6 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.jsonArray
-import kotlinx.serialization.json.jsonObject
-import kotlinx.serialization.json.jsonPrimitive
 import org.cashu.wallet.Core.CDK.CdkWalletGateway
 import org.cashu.wallet.Core.Platform.WalletDatabasePathManager
 import org.cashu.wallet.Core.Protocols.SecureStorage
@@ -26,35 +21,11 @@ import org.cashu.wallet.Models.MeltPaymentResult
 import org.cashu.wallet.Models.MeltQuoteInfo
 import org.cashu.wallet.Models.MintInfo
 import org.cashu.wallet.Models.MintQuoteInfo
-import org.cashu.wallet.Models.MintQuoteState
 import org.cashu.wallet.Models.PaymentMethodKind
-import org.cashu.wallet.Models.ClaimedToken
 import org.cashu.wallet.Models.PendingReceiveToken
 import org.cashu.wallet.Models.PendingToken
 import org.cashu.wallet.Models.RestoreMintResult
 import org.cashu.wallet.Models.SendTokenResult
-import org.cashu.wallet.Models.TransactionKind
-import org.cashu.wallet.Models.TransactionStatus
-import org.cashu.wallet.Models.TransactionType
-import org.cashu.wallet.Models.WalletTransaction
-
-data class WalletState(
-    val balance: Long = 0,
-    val pendingBalance: Long = 0,
-    val isInitialized: Boolean = false,
-    val needsOnboarding: Boolean = true,
-    val canExitOnboarding: Boolean = false,
-    val isLoading: Boolean = false,
-    val errorMessage: String? = null,
-    val activeUnit: String = "sat",
-    val mints: List<MintInfo> = emptyList(),
-    val activeMint: MintInfo? = null,
-    val transactions: List<WalletTransaction> = emptyList(),
-    val pendingTokens: List<PendingToken> = emptyList(),
-    val pendingReceiveTokens: List<PendingReceiveToken> = emptyList(),
-    val claimedTokens: List<ClaimedToken> = emptyList(),
-    val transactionUpdateVersion: Long = 0,
-)
 
 class WalletManager(
     private val secureStorage: SecureStorage,
@@ -72,7 +43,9 @@ class WalletManager(
     private val scope = CoroutineScope(SupervisorJob() + Dispatchers.Main.immediate + exceptionHandler)
     private val mutableState = MutableStateFlow(WalletState())
     val state: StateFlow<WalletState> = mutableState.asStateFlow()
-    private val mintQuoteSyncsInFlight = mutableSetOf<String>()
+    private val mintMetadataFetcher = WalletMintMetadataFetcher()
+    private val mintQuoteSyncService = WalletMintQuoteSyncService(gateway, walletStore)
+    private val transactionLoader = WalletTransactionLoader(walletStore, gateway)
     private val npcQuotesInFlight = mutableSetOf<String>()
     private var processedNPCQuotes = walletStore.loadProcessedNPCQuotes().toMutableSet()
 
@@ -177,14 +150,14 @@ class WalletManager(
 
     override suspend fun addMint(url: String) {
         withLoading {
-            val normalized = normalizeMintUrl(url)
-            validateMintUrl(normalized)?.let { throw IllegalArgumentException(it) }
+            val normalized = mintMetadataFetcher.normalizeMintUrl(url)
+            mintMetadataFetcher.validateMintUrl(normalized)?.let { throw IllegalArgumentException(it) }
             if (mutableState.value.mints.any { it.url == normalized }) {
                 throw IllegalArgumentException("Mint already exists.")
             }
             runCatching { gateway.ensureWallet(normalized) }
                 .onFailure { AppLogger.wallet.error("CDK wallet preparation is not available yet for $normalized", it) }
-            val fetched = gateway.fetchMintInfo(normalized) ?: fetchRawMintInfo(normalized)
+            val fetched = gateway.fetchMintInfo(normalized) ?: mintMetadataFetcher.fetchRawMintInfo(normalized)
             val updated = mutableState.value.mints + fetched
             walletStore.saveMints(updated)
             if (mutableState.value.activeMint == null) walletStore.activeMintURL = fetched.url
@@ -214,8 +187,8 @@ class WalletManager(
 
     override suspend fun restoreFromMint(url: String): RestoreMintResult =
         withLoadingResult {
-            val normalized = normalizeMintUrl(url)
-            validateMintUrl(normalized)?.let { throw IllegalArgumentException(it) }
+            val normalized = mintMetadataFetcher.normalizeMintUrl(url)
+            mintMetadataFetcher.validateMintUrl(normalized)?.let { throw IllegalArgumentException(it) }
             val trackedMintUrl = ensureMintTracked(normalized)
             val result = withContext(Dispatchers.IO) { gateway.restoreMint(trackedMintUrl) }
             refreshBalance()
@@ -244,15 +217,23 @@ class WalletManager(
     override suspend fun createMintQuote(amount: Long?, method: PaymentMethodKind): MintQuoteInfo {
         val active = mutableState.value.activeMint ?: throw IllegalStateException("No active mint.")
         return withLoadingResult {
-            gateway.createMintQuote(amount, method, active.url).also { rememberMintQuoteTimestamp(it.id) }
+            gateway.createMintQuote(amount, method, active.url).also {
+                mintQuoteSyncService.rememberMintQuoteTimestamp(it.id)
+            }
         }
     }
 
     suspend fun checkMintQuote(quoteId: String): MintQuoteInfo =
-        withLoadingResult { gateway.checkMintQuote(quoteId).also { rememberMintQuoteTimestamp(it.id) } }
+        withLoadingResult {
+            gateway.checkMintQuote(quoteId).also {
+                mintQuoteSyncService.rememberMintQuoteTimestamp(it.id)
+            }
+        }
 
     suspend fun pollMintQuote(quoteId: String): MintQuoteInfo =
-        gateway.checkMintQuote(quoteId).also { rememberMintQuoteTimestamp(it.id) }
+        gateway.checkMintQuote(quoteId).also {
+            mintQuoteSyncService.rememberMintQuoteTimestamp(it.id)
+        }
 
     fun subscribeToMintQuote(quoteId: String): Flow<MintQuoteInfo> = gateway.subscribeToMintQuote(quoteId)
 
@@ -266,7 +247,10 @@ class WalletManager(
 
     suspend fun refreshPendingMintQuote(quoteId: String): Boolean =
         withLoadingResult {
-            val minted = syncPendingMintQuote(quoteId, allowPendingOnchainMintAttempt = true)
+            val minted = mintQuoteSyncService.syncPendingMintQuote(
+                quoteId,
+                allowPendingOnchainMintAttempt = true,
+            )
             if (minted) refreshBalance()
             loadTransactions()
             minted
@@ -278,7 +262,12 @@ class WalletManager(
                 .getOrDefault(emptyList())
             var mintedCount = 0
             pendingQuotes.forEach { quote ->
-                if (syncPendingMintQuote(quote.id, allowPendingOnchainMintAttempt = false)) {
+                if (
+                    mintQuoteSyncService.syncPendingMintQuote(
+                        quote.id,
+                        allowPendingOnchainMintAttempt = false,
+                    )
+                ) {
                     mintedCount += 1
                 }
             }
@@ -304,7 +293,7 @@ class WalletManager(
             loadTransactions()
             amount > 0 || isNPCQuoteProcessed(quote.id)
         } catch (error: Throwable) {
-            if (isAlreadyIssuedMintError(error)) {
+            if (mintQuoteSyncService.isAlreadyIssuedMintError(error)) {
                 markNPCQuoteProcessed(quote.id)
                 true
             } else {
@@ -322,7 +311,7 @@ class WalletManager(
     override suspend fun meltTokens(quoteId: String, mintUrl: String?): MeltPaymentResult =
         withLoadingResult {
             val result = gateway.meltTokens(quoteId, mintUrl)
-            saveMeltPaymentMetadata(quoteId, result)
+            transactionLoader.saveMeltPaymentMetadata(quoteId, result)
             refreshBalance()
             loadTransactions()
             result
@@ -402,7 +391,13 @@ class WalletManager(
         withLoadingResult {
             val claimed = gateway.checkTokenSpendable(pendingToken.token, pendingToken.mintUrl)
             if (claimed) {
-                markPendingTokenClaimed(pendingToken)
+                val mutation = transactionLoader.markPendingTokenClaimed(pendingToken)
+                update {
+                    copy(
+                        pendingTokens = mutation.pendingTokens,
+                        claimedTokens = mutation.claimedTokens,
+                    )
+                }
                 loadTransactions()
             }
             claimed
@@ -416,7 +411,13 @@ class WalletManager(
                 val claimed = runCatching { gateway.checkTokenSpendable(token.token, token.mintUrl) }
                     .getOrDefault(false)
                 if (claimed) {
-                    markPendingTokenClaimed(token)
+                    val mutation = transactionLoader.markPendingTokenClaimed(token)
+                    update {
+                        copy(
+                            pendingTokens = mutation.pendingTokens,
+                            claimedTokens = mutation.claimedTokens,
+                        )
+                    }
                     claimedCount += 1
                 }
             }
@@ -446,60 +447,13 @@ class WalletManager(
 
     suspend fun loadTransactions() {
         val mintUrls = mutableState.value.mints.map { it.url }
-        val trackedMintUrls = mintUrls.toSet()
-        val preimages = walletStore.loadPaymentPreimages()
-        val meltFees = walletStore.loadMeltQuoteFees()
-        val pendingTokens = walletStore.loadPendingTokens()
-        val pendingReceiveTokens = walletStore.loadPendingReceiveTokens()
-        val claimedTokens = walletStore.loadClaimedTokens()
-        val remote = runCatching { gateway.listTransactions(mintUrls) }.getOrDefault(emptyList())
-            .map { it.withStoredMeltMetadata(preimages, meltFees) }
-        val completedQuoteIds = remote.mapNotNull { it.quoteId }.toSet()
-        val mintQuoteTimestamps = walletStore.loadMintQuoteTimestamps().toMutableMap()
-        val pendingQuoteTransactions = observePendingOnchainMintQuotes(
-            runCatching { gateway.listUnissuedMintQuotes() }
-                .getOrDefault(emptyList())
-                .let { quotes ->
-                    pendingMintQuoteTransactions(
-                        quotes = quotes,
-                        trackedMintUrls = trackedMintUrls,
-                        completedQuoteIds = completedQuoteIds,
-                        timestamps = mintQuoteTimestamps,
-                        nowEpochMillis = System.currentTimeMillis(),
-                    )
-                }
-                .map { it.withStoredMeltMetadata(preimages, meltFees) },
-        )
-        val storedMeltTransactions = runCatching { gateway.listMeltQuotes() }
-            .getOrDefault(emptyList())
-            .let { quotes ->
-                storedMeltQuoteTransactions(
-                    quotes = quotes,
-                    trackedMintUrls = trackedMintUrls,
-                    completedQuoteIds = completedQuoteIds,
-                    timestamps = mintQuoteTimestamps,
-                    nowEpochMillis = System.currentTimeMillis(),
-                    preimages = preimages,
-                    fees = meltFees,
-                )
-            }
-        val tokenTransactions = pendingSentTokenTransactions(pendingTokens) +
-            pendingReceiveTokenTransactions(pendingReceiveTokens) +
-            claimedTokenTransactions(claimedTokens)
-        val cached = walletStore.loadTransactions()
-            .filterNot { it.isPendingToken }
-            .map { it.withStoredMeltMetadata(preimages, meltFees) }
-        val merged = (remote + pendingQuoteTransactions + storedMeltTransactions + tokenTransactions + cached)
-            .distinctBy { "${it.mintUrl.orEmpty()}|${it.quoteId ?: it.id}" }
-            .sortedByDescending { it.dateEpochMillis }
-        walletStore.saveTransactions(merged)
-        walletStore.saveMintQuoteTimestamps(pruneMintQuoteTimestamps(merged, mintQuoteTimestamps))
+        val result = transactionLoader.load(mintUrls)
         update {
             copy(
-                transactions = merged,
-                pendingTokens = pendingTokens,
-                pendingReceiveTokens = pendingReceiveTokens,
-                claimedTokens = claimedTokens,
+                transactions = result.transactions,
+                pendingTokens = result.pendingTokens,
+                pendingReceiveTokens = result.pendingReceiveTokens,
+                claimedTokens = result.claimedTokens,
                 transactionUpdateVersion = nextTransactionUpdateVersion(transactionUpdateVersion),
             )
         }
@@ -598,171 +552,9 @@ class WalletManager(
         }
     }
 
-    private fun markPendingTokenClaimed(pendingToken: PendingToken) {
-        val pending = walletStore.loadPendingTokens().filterNot { it.tokenId == pendingToken.tokenId }
-        val claimedToken = ClaimedToken(
-            tokenId = pendingToken.tokenId,
-            token = pendingToken.token,
-            amount = pendingToken.amount,
-            fee = pendingToken.fee,
-            dateEpochMillis = pendingToken.dateEpochMillis,
-            mintUrl = pendingToken.mintUrl,
-            memo = pendingToken.memo,
-            claimedDateEpochMillis = System.currentTimeMillis(),
-        )
-        val claimed = walletStore.loadClaimedTokens()
-            .filterNot { it.tokenId == pendingToken.tokenId } + claimedToken
-        walletStore.savePendingTokens(pending)
-        walletStore.saveClaimedTokens(claimed)
-        update { copy(pendingTokens = pending, claimedTokens = claimed) }
-    }
-
     private fun markNPCQuoteProcessed(quoteId: String) {
         processedNPCQuotes += quoteId
         walletStore.saveProcessedNPCQuotes(processedNPCQuotes.sorted())
-    }
-
-    private fun saveMeltPaymentMetadata(quoteId: String, result: MeltPaymentResult) {
-        result.preimage?.let { preimage ->
-            walletStore.savePaymentPreimages(walletStore.loadPaymentPreimages() + (quoteId to preimage))
-        }
-        walletStore.saveMeltQuoteFees(walletStore.loadMeltQuoteFees() + (quoteId to result.feePaid))
-
-        val current = walletStore.loadTransactions()
-        val existing = current.firstOrNull { it.quoteId == quoteId || it.id == quoteId }
-        val transaction = WalletTransaction(
-            id = existing?.id ?: quoteId,
-            amount = result.amount,
-            type = TransactionType.Outgoing,
-            kind = when (result.paymentMethod) {
-                PaymentMethodKind.Onchain -> TransactionKind.Onchain
-                else -> TransactionKind.Lightning
-            },
-            dateEpochMillis = existing?.dateEpochMillis ?: System.currentTimeMillis(),
-            memo = existing?.memo,
-            status = TransactionStatus.Completed,
-            mintUrl = result.mintUrl,
-            preimage = result.preimage ?: existing?.preimage,
-            invoice = result.request ?: existing?.invoice,
-            fee = result.feePaid,
-            quoteId = quoteId,
-        )
-        walletStore.saveTransactions(
-            listOf(transaction) + current.filterNot { it.id == transaction.id || it.quoteId == quoteId },
-        )
-    }
-
-    private suspend fun syncPendingMintQuote(
-        quoteId: String,
-        allowPendingOnchainMintAttempt: Boolean,
-    ): Boolean {
-        if (!mintQuoteSyncsInFlight.add(quoteId)) return false
-        return try {
-            val updatedQuote = gateway.checkMintQuote(quoteId).also { rememberMintQuoteTimestamp(it.id) }
-            val shouldAttemptMint = updatedQuote.state == MintQuoteState.Paid ||
-                updatedQuote.state == MintQuoteState.Issued ||
-                (allowPendingOnchainMintAttempt && updatedQuote.paymentMethod == PaymentMethodKind.Onchain)
-            if (!shouldAttemptMint) return false
-
-            if (updatedQuote.paymentMethod == PaymentMethodKind.Bolt12 &&
-                updatedQuote.amountPaid > 0 &&
-                updatedQuote.amountIssued >= updatedQuote.amountPaid
-            ) {
-                return false
-            }
-
-            runCatching { gateway.mintTokens(quoteId) }
-                .fold(
-                    onSuccess = { true },
-                    onFailure = { error ->
-                        if (isAlreadyIssuedMintError(error)) {
-                            true
-                        } else if (
-                            updatedQuote.paymentMethod == PaymentMethodKind.Onchain &&
-                            updatedQuote.state == MintQuoteState.Pending
-                        ) {
-                            false
-                        } else {
-                            AppLogger.wallet.error("Failed to mint pending quote $quoteId", error)
-                            false
-                        }
-                    },
-                )
-        } catch (error: Throwable) {
-            if (!isMissingQuoteError(error)) {
-                AppLogger.wallet.error("Failed to refresh pending quote $quoteId", error)
-            }
-            false
-        } finally {
-            mintQuoteSyncsInFlight.remove(quoteId)
-        }
-    }
-
-    private suspend fun observePendingOnchainMintQuotes(
-        transactions: List<WalletTransaction>,
-    ): List<WalletTransaction> =
-        transactions.map { transaction ->
-            if (
-                transaction.type != TransactionType.Incoming ||
-                transaction.kind != TransactionKind.Onchain ||
-                transaction.invoice == null
-            ) {
-                return@map transaction
-            }
-
-            val observation = OnchainExplorer.observePayment(
-                address = transaction.invoice,
-                mintUrl = transaction.mintUrl,
-                expectedAmount = transaction.amount,
-                createdAfterEpochMillis = transaction.dateEpochMillis,
-            )
-
-            if (observation != null) {
-                val key = transaction.quoteId ?: transaction.id
-                val currentPreimages = walletStore.loadPaymentPreimages()
-                if (currentPreimages[key] != observation.txid) {
-                    walletStore.savePaymentPreimages(currentPreimages + (key to observation.txid))
-                }
-                transaction.copy(
-                    preimage = observation.txid,
-                    statusNote = observation.statusText,
-                )
-            } else if (transaction.preimage != null) {
-                transaction.copy(statusNote = transaction.statusNote ?: "Payment detected on-chain")
-            } else {
-                transaction
-            }
-        }
-
-    private fun rememberMintQuoteTimestamp(quoteId: String) {
-        val current = walletStore.loadMintQuoteTimestamps()
-        if (quoteId !in current) {
-            walletStore.saveMintQuoteTimestamps(current + (quoteId to System.currentTimeMillis()))
-        }
-    }
-
-    private fun isMissingQuoteError(error: Throwable): Boolean {
-        val message = "${error.message.orEmpty()} ${error}".lowercase()
-        return message.contains("not found") ||
-            message.contains("no stored mint quote") ||
-            message.contains("missing quote")
-    }
-
-    private fun isAlreadyIssuedMintError(error: Throwable): Boolean {
-        val message = "${error.message.orEmpty()} ${error}".lowercase()
-        if (
-            message.contains("already being minted") ||
-            message.contains("not issued") ||
-            message.contains("not yet") ||
-            message.contains("unissued")
-        ) {
-            return false
-        }
-        return message.contains("already issued") ||
-            message.contains("already minted") ||
-            message.contains("quote is issued") ||
-            message.contains("state=issued") ||
-            message.contains("tokens already issued")
     }
 
     private fun activeMintFrom(mints: List<MintInfo>): MintInfo? {
@@ -771,14 +563,14 @@ class WalletManager(
     }
 
     private suspend fun ensureMintTracked(url: String): String {
-        val normalized = normalizeMintUrl(url)
+        val normalized = mintMetadataFetcher.normalizeMintUrl(url)
         runCatching { gateway.ensureWallet(normalized) }
             .onFailure { AppLogger.wallet.error("CDK wallet preparation is not available yet for $normalized", it) }
         if (walletStore.loadMints().any { it.url == normalized }) return normalized
 
         val fetched = runCatching { gateway.fetchMintInfo(normalized) }
             .getOrNull()
-            ?: runCatching { fetchRawMintInfo(normalized) }.getOrElse {
+            ?: runCatching { mintMetadataFetcher.fetchRawMintInfo(normalized) }.getOrElse {
                 MintInfo(
                     url = normalized,
                     name = runCatching { URL(normalized).host }.getOrNull() ?: "Unknown Mint",
@@ -795,55 +587,6 @@ class WalletManager(
             )
         }
         return normalized
-    }
-
-    private suspend fun fetchRawMintInfo(url: String): MintInfo = withContext(Dispatchers.IO) {
-        val connection = (URL("$url/v1/info").openConnection() as HttpURLConnection).apply {
-            requestMethod = "GET"
-            connectTimeout = 10_000
-            readTimeout = 10_000
-        }
-        try {
-            if (connection.responseCode !in 200..299) throw IllegalStateException("Mint info HTTP ${connection.responseCode}")
-            val body = connection.inputStream.bufferedReader().use { it.readText() }
-            val root = Json.parseToJsonElement(body).jsonObject
-            val name = root["name"]?.jsonPrimitive?.content ?: URL(url).host ?: "Unknown Mint"
-            val description = root["description"]?.jsonPrimitive?.content
-            val iconUrl = root["icon_url"]?.jsonPrimitive?.content
-            val nuts = root["nuts"]?.jsonObject
-            val nut04 = nuts?.get("4")?.jsonObject
-            val methods = nut04?.get("methods")?.jsonArray.orEmpty()
-            val supportsOnchain = methods.any { element ->
-                element.jsonObject["method"]?.jsonPrimitive?.content?.lowercase() == "onchain"
-            }
-            MintInfo(
-                url = url,
-                name = name,
-                description = description,
-                iconUrl = iconUrl,
-                supportedMintMethods = listOfNotNull(
-                    PaymentMethodKind.Bolt11,
-                    PaymentMethodKind.Onchain.takeIf { supportsOnchain },
-                ),
-            )
-        } finally {
-            connection.disconnect()
-        }
-    }
-
-    private fun normalizeMintUrl(url: String): String {
-        var normalized = url.trim()
-        if (!normalized.startsWith("http://") && !normalized.startsWith("https://")) {
-            normalized = "https://$normalized"
-        }
-        return normalized.trimEnd('/')
-    }
-
-    private fun validateMintUrl(url: String): String? {
-        val parsed = runCatching { URL(url) }.getOrNull() ?: return "Invalid URL format."
-        if (parsed.host.isNullOrBlank()) return "Invalid URL format."
-        if (parsed.protocol != "https") return "Mint URL must use HTTPS for security."
-        return null
     }
 
     private suspend fun deriveNostrKey(mnemonic: String) {
@@ -887,24 +630,4 @@ class WalletManager(
     fun reopenOnboarding() {
         update { copy(needsOnboarding = true, canExitOnboarding = true) }
     }
-}
-
-internal fun WalletTransaction.withStoredMeltMetadata(
-    preimages: Map<String, String>,
-    meltFees: Map<String, Long>,
-): WalletTransaction {
-    val key = quoteId ?: id
-    return copy(
-        preimage = preimage ?: preimages[key],
-        fee = meltFees[key] ?: fee,
-    )
-}
-
-internal fun shouldAttemptWalletDatabaseRecovery(error: Throwable): Boolean {
-    val normalized = (error.message ?: error.toString()).lowercase()
-    return normalized.contains("sqlite") ||
-        normalized.contains("database") ||
-        normalized.contains("corrupt") ||
-        normalized.contains("malformed") ||
-        normalized.contains("walletdb")
 }
