+++ android/app/src/main/java/org/cashu/wallet/Models/Payments/OnchainPaymentObservation.kt
@@ -1,0 +1,21 @@
+package org.cashu.wallet.Models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class OnchainPaymentObservation(
+    val txid: String,
+    val amount: Long,
+    val confirmed: Boolean,
+    val confirmations: Int? = null,
+) {
+    val statusText: String
+        get() = when {
+            confirmations != null && confirmations > 0 -> {
+                val suffix = if (confirmations == 1) "" else "s"
+                "Payment confirmed on-chain ($confirmations confirmation$suffix)"
+            }
+            confirmed -> "Payment detected on-chain"
+            else -> "Payment seen in mempool"
+        }
+}
