+++ android/app/src/main/java/org/cashu/wallet/Models/Settings/NwcConnection.kt
@@ -1,0 +1,13 @@
+package org.cashu.wallet.Models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class NwcConnection(
+    val id: String,
+    val name: String,
+    val walletPublicKey: String,
+    val connectionPublicKey: String,
+    val allowanceSats: Long?,
+    val createdAtEpochMillis: Long = System.currentTimeMillis(),
+)
