+++ android/app/src/main/java/org/cashu/wallet/Models/Tokens/TokenInfo.kt
@@ -1,0 +1,17 @@
+package org.cashu.wallet.Models
+
+import kotlinx.serialization.Serializable
+import org.cashu.wallet.Core.TokenParser
+
+@Serializable
+data class TokenInfo(
+    val amount: Long,
+    val mint: String,
+    val unit: String,
+    val memo: String?,
+    val proofCount: Int,
+) {
+    companion object {
+        fun parse(tokenString: String): TokenInfo? = TokenParser.tokenInfo(tokenString)
+    }
+}
